#68 made navbar CSR
